Organise files in file-viewer directory
