+++ app/components/file-viewer/text-viewer.jsx
